init depth container
This is intended to prepare one container for depth 16bit data.
As we know depth data is becoming more and more important these days,
especially for the Face-Antispoofing.
We are thinking of using some could service with depth data operation,
but we have to use 8bit compressed data, but with less precision.
After some discussion, our team member and our mentors/partners
valued the 16bit depth data.
The first step is to record some depth raw data. And we don't even have
a container. So I created this repo hoping to create some container tool
for better investigation/try out.

Signed-off-by: Juan Zhao <juan.j.zhao@intel.com>